implemented overrides for Box

+++ include/element/Box.hpp
@@ -30,7 +30,7 @@ namespace reng
         /** Increments the width and height by given values */
         void scale(int dWidth, int dHeight);
         /** Scales the image by a multiplier */
-        void scale(int multiplier);
+        void scale(double multiplier);
     };
 }
